# src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Neo Domotique API",
    "description": "API Backend pour l'application Neo Domotique - Solution de domotique white-label",
    "version": "1.0.0",
    "contact": {
      "name": "Neo Domotique",
      "email": "contact@neo-domotique.fr"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Connexion",
        "description": "Authentifie un utilisateur et retourne les tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "jean.dupont@neo-domotique.fr" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Déconnexion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Déconnexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Déconnexion réussie" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rafraîchir le token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nouveaux tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Utilisateur courant",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Informations utilisateur",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Liste des utilisateurs",
        "description": "Admin only",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "integrateur", "auditeur"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedUsers" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Créer un utilisateur",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUser" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Détail utilisateur",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Modifier utilisateur",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUser" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Supprimer utilisateur",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Deleted" }
        }
      }
    },
...
    "/api/projets/{projectId}/pieces": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Liste des pièces d'un projet",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Room" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Rooms"],
        "summary": "Créer une pièce",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRoom" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Room" }
              }
            }
          }
        }
      }
    },
    "/api/pieces/{id}": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Détail pièce",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomDetail" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Rooms"],
        "summary": "Modifier pièce",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateRoom" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Room" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Rooms"],
        "summary": "Supprimer pièce",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Deleted" }
        }
      }
    },
...
    "/api/pieces/{roomId}/checklist": {
      "post": {
        "tags": ["Checklist"],
        "summary": "Créer un item checklist",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RoomIdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateChecklistItem" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChecklistItem" }
              }
            }
          }
        }
      }
    },
    "/api/checklist/{id}": {
      "put": {
        "tags": ["Checklist"],
        "summary": "Modifier item checklist",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateChecklistItem" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChecklistItem" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Checklist"],
        "summary": "Supprimer item checklist",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Deleted" }
        }
      }
    },
...
    "/api/sync/status": {
      "get": {
        "tags": ["Sync"],
        "summary": "Statut de synchronisation",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lastSync": { "type": "string", "format": "date-time" },
                    "pendingChanges": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/pull": {
      "post": {
        "tags": ["Sync"],
        "summary": "Pull changes from server",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lastSyncAt": { "type": "string", "format": "date-time" },
                  "tables": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["clients", "projects", "rooms"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "changes": { "type": "object" },
                    "serverTimestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/push": {
      "post": {
        "tags": ["Sync"],
        "summary": "Push changes to server",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "changes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "tableName": { "type": "string" },
                        "recordId": { "type": "string" },
                        "operation": { "type": "string", "enum": ["create", "update", "delete"] },
                        "data": { "type": "object" },
                        "clientTimestamp": { "type": "string", "format": "date-time" }
                      }
                    }
                  },
                  "deviceId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "processed": { "type": "integer" },
                    "errors": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "ProjectIdPath": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "RoomIdPath": {
        "name": "roomId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 1, "minimum": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Non authentifié",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "Deleted": {
        "description": "Supprimé",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Supprimé" }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "code": { "type": "string" }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "integrateur", "auditeur"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateUser": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "integrateur", "auditeur"], "default": "integrateur" }
        }
      },
      "UpdateUser": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "integrateur", "auditeur"] }
        }
      },
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
...
      "Room": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "projectId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["salon", "cuisine", "chambre", "salle_de_bain", "bureau", "garage", "exterieur", "autre"] },
          "floor": { "type": "integer" },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "RoomDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Room" },
          {
            "type": "object",
            "properties": {
              "photos": { "type": "array", "items": { "$ref": "#/components/schemas/Photo" } },
              "checklistItems": { "type": "array", "items": { "$ref": "#/components/schemas/ChecklistItem" } },
              "devices": { "type": "array", "items": { "$ref": "#/components/schemas/Device" } }
            }
          }
        ]
      },
      "CreateRoom": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["salon", "cuisine", "chambre", "salle_de_bain", "bureau", "garage", "exterieur", "autre"], "default": "autre" },
          "floor": { "type": "integer", "default": 0 },
          "notes": { "type": "string" }
        }
      },
      "UpdateRoom": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["salon", "cuisine", "chambre", "salle_de_bain", "bureau", "garage", "exterieur", "autre"] },
          "floor": { "type": "integer" },
          "notes": { "type": "string" }
        }
      },
      "Device": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "roomId": { "type": "string", "format": "uuid" },
          "productId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "serialNumber": { "type": "string" },
          "macAddress": { "type": "string" },
          "ipAddress": { "type": "string" },
          "status": { "type": "string", "enum": ["planifie", "installe", "configure", "operationnel", "en_panne"] },
          "location": { "type": "string" },
          "notes": { "type": "string" },
          "isOnline": { "type": "boolean" },
          "lastSeenAt": { "type": "string", "format": "date-time" },
          "installedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Photo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "roomId": { "type": "string", "format": "uuid" },
          "filename": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "caption": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ChecklistItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "roomId": { "type": "string", "format": "uuid" },
          "category": { "type": "string" },
          "label": { "type": "string" },
          "checked": { "type": "boolean" },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateChecklistItem": {
        "type": "object",
        "required": ["category", "label"],
        "properties": {
          "category": { "type": "string" },
          "label": { "type": "string" },
          "checked": { "type": "boolean", "default": false },
          "notes": { "type": "string" }
        }
      },
      "UpdateChecklistItem": {
        "type": "object",
        "properties": {
          "category": { "type": "string" },
          "label": { "type": "string" },
          "checked": { "type": "boolean" },
          "notes": { "type": "string" }
        }
      },
...
    }
  }
}